fix format

+++ src/SimpleBank.sol
@@ -3,7 +3,6 @@
 pragma solidity ^0.8.30;
 
 contract SimpleBank {
-
     mapping(address => uint256) public userBalance;
 
     function deposit() public payable {
@@ -16,13 +15,13 @@ contract SimpleBank {
         require(userBalance[msg.sender] >= 1 ether, "User has not enough balance");
         require(address(this).balance > 0, "Ban is rekt");
 
-        (bool success, ) = msg.sender.call{value: userBalance[msg.sender]}("");
+        (bool success,) = msg.sender.call{value: userBalance[msg.sender]}("");
         require(success, "fail");
 
         userBalance[msg.sender] = 0;
     }
 
-    function totalBalance() public view returns(uint) {
+    function totalBalance() public view returns (uint256) {
         return address(this).balance;
     }
 }
